Updated file structure
The file structure was a mess so it was updated.
Subject to change.
