+++ src/Attachment.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "base/Attachment.h"
+#include "Attachment.h"
 
 using namespace GroupMe;
