+++ src/File.cpp
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "base/File.h"
+#include "File.h"
 
 using namespace GroupMe;
