+++ src/Message.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "base/Message.h"
+#include "Message.h"
 
 using namespace GroupMe;
