+++ src/Picture.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "base/Picture.h"
+#include "Picture.h"
 
 using namespace GroupMe;
