+++ src/User.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "base/User.h"
+#include "User.h"
 
 using namespace GroupMe;
