+++ src/Video.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "base/Video.h"
+#include "Video.h"
 
 using namespace GroupMe;
