+++ src/util/AVFileMem.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include <base/AVFileMem.h>
+#include <util/AVFileMem.h>
 
 using namespace GroupMe::Util;
